Fix OpenVPN import issue
Previously, the command `htb lab import` would fail.  This is because it
was writing an empty file, as the write() would be buffered and not
flush to file before the `nmcli` import was called.

This change explicitly calls flush() to commit the changes to the file,
meaning the contents is written to the file.

Fixes: calebstewart/python-htb#6
Signed-off-by: Dave Walker (Daviey) <email@daviey.com>